Move tests so they are not considered part of flow
