+++ flow-server/src/test/java/com/vaadin/notreallyflow/ComponentTrackerTest.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow;
+package com.vaadin.notreallyflow;
 
 import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.HasComponents;
